a little refactor and unit tests for A/B testing methods
