+++ electrode-ota-server-routes-acquisition/yarn-error.log
@@ -1,0 +1,153 @@
+Arguments: 
+  /Users/c0s03o1/.nvm/versions/node/v10.16.0/bin/node /Users/c0s03o1/.nvm/versions/node/v10.16.0/bin/yarn install --mutex network:42424 --non-interactive
+
+PATH: 
+  /Users/c0s03o1/Documents/Projects/Workspace/ota/electrode-ota-server/electrode-ota-server-routes-acquisition/node_modules/.bin:/Users/c0s03o1/Documents/Projects/Workspace/ota/electrode-ota-server/node_modules/.bin:/Users/c0s03o1/Documents/Projects/Workspace/ota/node_modules/.bin:/Users/c0s03o1/Documents/Projects/Workspace/node_modules/.bin:/Users/c0s03o1/Documents/Projects/node_modules/.bin:/Users/c0s03o1/Documents/node_modules/.bin:/Users/c0s03o1/node_modules/.bin:/Users/node_modules/.bin:/node_modules/.bin:/Users/c0s03o1/.nvm/versions/node/v10.16.0/bin:/Users/c0s03o1/.nvm/versions/node/v10.16.0/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin
+
+Yarn version: 
+  1.22.4
+
+Node version: 
+  10.16.0
+
+Platform: 
+  darwin x64
+
+Trace: 
+  Error: unable to get local issuer certificate
+      at TLSSocket.onConnectSecure (_tls_wrap.js:1055:34)
+      at TLSSocket.emit (events.js:198:13)
+      at TLSSocket._finishInit (_tls_wrap.js:633:8)
+
+npm manifest: 
+  {
+  	"name": "electrode-ota-server-routes-acquisition",
+  	"version": "4.6.10-beta.0",
+  	"main": "lib/index.js",
+  	"devDependencies": {
+  		"proxyquire": "2.1.3",
+  		"sinon": "7.4.1"
+  	}
+  }
+
+yarn manifest: 
+  No manifest
+
+Lockfile: 
+  # THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.
+  # yarn lockfile v1
+  
+  
+  "@sinonjs/commons@^1", "@sinonjs/commons@^1.3.0", "@sinonjs/commons@^1.4.0", "@sinonjs/commons@^1.7.0":
+    version "1.7.2"
+    resolved "https://registry.yarnpkg.com/@sinonjs/commons/-/commons-1.7.2.tgz#505f55c74e0272b43f6c52d81946bed7058fc0e2"
+    integrity sha512-+DUO6pnp3udV/v2VfUWgaY5BIE1IfT7lLfeDzPVeMT1XKkaAp9LgSI9x5RtrFQoZ9Oi0PgXQQHPaoKu7dCjVxw==
+    dependencies:
+      type-detect "4.0.8"
+  
+  "@sinonjs/formatio@^3.2.1":
+    version "3.2.2"
+    resolved "https://registry.yarnpkg.com/@sinonjs/formatio/-/formatio-3.2.2.tgz#771c60dfa75ea7f2d68e3b94c7e888a78781372c"
+    integrity sha512-B8SEsgd8gArBLMD6zpRw3juQ2FVSsmdd7qlevyDqzS9WTCtvF55/gAL+h6gue8ZvPYcdiPdvueM/qm//9XzyTQ==
+    dependencies:
+      "@sinonjs/commons" "^1"
+      "@sinonjs/samsam" "^3.1.0"
+  
+  "@sinonjs/samsam@^3.1.0", "@sinonjs/samsam@^3.3.2":
+    version "3.3.3"
+    resolved "https://registry.yarnpkg.com/@sinonjs/samsam/-/samsam-3.3.3.tgz#46682efd9967b259b81136b9f120fd54585feb4a"
+    integrity sha512-bKCMKZvWIjYD0BLGnNrxVuw4dkWCYsLqFOUWw8VgKF/+5Y+mE7LfHWPIYoDXowH+3a9LsWDMo0uAP8YDosPvHQ==
+    dependencies:
+      "@sinonjs/commons" "^1.3.0"
+      array-from "^2.1.1"
+      lodash "^4.17.15"
+  
+  "@sinonjs/text-encoding@^0.7.1":
+    version "0.7.1"
+    resolved "https://registry.yarnpkg.com/@sinonjs/text-encoding/-/text-encoding-0.7.1.tgz#8da5c6530915653f3a1f38fd5f101d8c3f8079c5"
+    integrity sha512-+iTbntw2IZPb/anVDbypzfQa+ay64MW0Zo8aJ8gZPWMMK6/OubMVb6lUPMagqjOPnmtauXnFCACVl3O7ogjeqQ==
+  
+  array-from@^2.1.1:
+    version "2.1.1"
+    resolved "https://registry.yarnpkg.com/array-from/-/array-from-2.1.1.tgz#cfe9d8c26628b9dc5aecc62a9f5d8f1f352c1195"
+    integrity sha1-z+nYwmYoudxa7MYqn12PHzUsEZU=
+  
+  diff@^3.5.0:
+    version "3.5.0"
+    resolved "https://registry.yarnpkg.com/diff/-/diff-3.5.0.tgz#800c0dd1e0a8bfbc95835c202ad220fe317e5a12"
+    integrity sha512-A46qtFgd+g7pDZinpnwiRJtxbC1hpgf0uzP3iG89scHk0AUC7A1TGxf5OiiOUv/JMZR8GOt8hL900hV0bOy5xA==
+  
+  has-flag@^3.0.0:
+    version "3.0.0"
+    resolved "https://registry.yarnpkg.com/has-flag/-/has-flag-3.0.0.tgz#b5d454dc2199ae225699f3467e5a07f3b955bafd"
+    integrity sha1-tdRU3CGZriJWmfNGfloH87lVuv0=
+  
+  isarray@0.0.1:
+    version "0.0.1"
+    resolved "https://registry.yarnpkg.com/isarray/-/isarray-0.0.1.tgz#8a18acfca9a8f4177e09abfc6038939b05d1eedf"
+    integrity sha1-ihis/Kmo9Bd+Cav8YDiTmwXR7t8=
+  
+  just-extend@^4.0.2:
+    version "4.1.0"
+    resolved "https://registry.yarnpkg.com/just-extend/-/just-extend-4.1.0.tgz#7278a4027d889601640ee0ce0e5a00b992467da4"
+    integrity sha512-ApcjaOdVTJ7y4r08xI5wIqpvwS48Q0PBG4DJROcEkH1f8MdAiNFyFxz3xoL0LWAVwjrwPYZdVHHxhRHcx/uGLA==
+  
+  lodash@^4.17.15:
+    version "4.17.15"
+    resolved "https://registry.yarnpkg.com/lodash/-/lodash-4.17.15.tgz#b447f6670a0455bbfeedd11392eff330ea097548"
+    integrity sha512-8xOcRHvCjnocdS5cpwXQXVzmmh5e5+saE2QGoeQmbKmRS6J3VQppPOIt0MnmE+4xlZoumy0GPG0D0MVIQbNA1A==
+  
+  lolex@^4.2.0:
+    version "4.2.0"
+    resolved "https://registry.yarnpkg.com/lolex/-/lolex-4.2.0.tgz#ddbd7f6213ca1ea5826901ab1222b65d714b3cd7"
+    integrity sha512-gKO5uExCXvSm6zbF562EvM+rd1kQDnB9AZBbiQVzf1ZmdDpxUSvpnAaVOP83N/31mRK8Ml8/VE8DMvsAZQ+7wg==
+  
+  lolex@^5.0.1:
+    version "5.1.2"
+    resolved "https://registry.yarnpkg.com/lolex/-/lolex-5.1.2.tgz#953694d098ce7c07bc5ed6d0e42bc6c0c6d5a367"
+    integrity sha512-h4hmjAvHTmd+25JSwrtTIuwbKdwg5NzZVRMLn9saij4SZaepCrTCxPr35H/3bjwfMJtN+t3CX8672UIkglz28A==
+    dependencies:
+      "@sinonjs/commons" "^1.7.0"
+  
+  nise@^1.5.1:
+    version "1.5.3"
+    resolved "https://registry.yarnpkg.com/nise/-/nise-1.5.3.tgz#9d2cfe37d44f57317766c6e9408a359c5d3ac1f7"
+    integrity sha512-Ymbac/94xeIrMf59REBPOv0thr+CJVFMhrlAkW/gjCIE58BGQdCj0x7KRCb3yz+Ga2Rz3E9XXSvUyyxqqhjQAQ==
+    dependencies:
+      "@sinonjs/formatio" "^3.2.1"
+      "@sinonjs/text-encoding" "^0.7.1"
+      just-extend "^4.0.2"
+      lolex "^5.0.1"
+      path-to-regexp "^1.7.0"
+  
+  path-to-regexp@^1.7.0:
+    version "1.8.0"
+    resolved "https://registry.yarnpkg.com/path-to-regexp/-/path-to-regexp-1.8.0.tgz#887b3ba9d84393e87a0a0b9f4cb756198b53548a"
+    integrity sha512-n43JRhlUKUAlibEJhPeir1ncUID16QnEjNpwzNdO3Lm4ywrBpBZ5oLD0I6br9evr1Y9JTqwRtAh7JLoOzAQdVA==
+    dependencies:
+      isarray "0.0.1"
+  
+  sinon@7.4.1:
+    version "7.4.1"
+    resolved "https://registry.yarnpkg.com/sinon/-/sinon-7.4.1.tgz#bcd0c63953893e87fa0cc502f52489c32a83d4d9"
+    integrity sha512-7s9buHGHN/jqoy/v4bJgmt0m1XEkCEd/tqdHXumpBp0JSujaT4Ng84JU5wDdK4E85ZMq78NuDe0I3NAqXY8TFg==
+    dependencies:
+      "@sinonjs/commons" "^1.4.0"
+      "@sinonjs/formatio" "^3.2.1"
+      "@sinonjs/samsam" "^3.3.2"
+      diff "^3.5.0"
+      lolex "^4.2.0"
+      nise "^1.5.1"
+      supports-color "^5.5.0"
+  
+  supports-color@^5.5.0:
+    version "5.5.0"
+    resolved "https://registry.yarnpkg.com/supports-color/-/supports-color-5.5.0.tgz#e2e69a44ac8772f78a1ec0b35b689df6530efc8f"
+    integrity sha512-QjVjwdXIt408MIiAqCX4oUKsgU2EqAGzs2Ppkm4aQYbjm+ZEWEcW4SfFNTr4uMNZma0ey4f5lgLrkB0aX0QMow==
+    dependencies:
+      has-flag "^3.0.0"
+  
+  type-detect@4.0.8:
+    version "4.0.8"
+    resolved "https://registry.yarnpkg.com/type-detect/-/type-detect-4.0.8.tgz#7646fb5f18871cfbb7749e69bd39a6388eb7450c"
+    integrity sha512-0fr/mIH1dlO+x7TlcMy+bIDqKPsw/70tVyeHW787goQjhmqaZe10uwLujubK9q9Lg6Fiho1KUKDYz0Z7k7g5/g==
